+++ vpr/Tables/GamePrediction.sql
@@ -5,7 +5,7 @@
     T1P2Rating    INT   NOT NULL,
     T2P1Rating    INT   NOT NULL,
     T2P2Rating    INT   NOT NULL,
-    T1PredictedWinProb  FLOAT NOT NULL,
+    T1PredictedWinProb  DECIMAL NOT NULL,
     ExpectT1Score INT   NULL,
     ExpectT2Score INT   NULL,
     CreatedAt     DATETIMEOFFSET NOT NULL , -- do not default, this is the app level concurrency token
